Drop 'licenses' property on data package general metadata
The licenses property with an array is unhelpful as it does not say
what license applies to what piece of work. The license property
can be either a string or a hash although there is an issue
discussing the possibility of only including a url which means it
could be further reduced to only the string version but that's
not what this commit tackles.