Drop 'licenses' property on resources in favour of 'license'
Using licenses as an array does not say which license applies to
what part of the resource. It is also very unlikely that a resource
will consist of work under different licenses rather than a single
one. Furthermore due to a previous commit, licenses are no longer
supported and data package only uses license.