Data Package: Reference CommonMark
The `description` is CommonMark by definition, so the word `MUST` is not
approriate. The first paragraph in a CommonMarkd document does not
necessarily end with two line breaks, for instance it can directly be
followed by one line break and a heading! This should better be a
recommendation to allow splitting it from the rest of the text without
use of a full-featured Markdown parser.